+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Changelog for aim.models
 
 - Services are loaded as entry_point plugins named `aim.services`
 
+- Refactored the models applications, resources, and services.
+
+- Added S3Resource for Resources/S3.yml configuration
+
 
 1.1.0 (2019-07-06)
 ------------------